[refactor] account app rename
