+++ eatda_be/settings/base.py
@@ -32,7 +32,7 @@ DEBUG = env('DEBUG')
 ALLOWED_HOSTS = ['*']
 
 # System User Default Model
-AUTH_USER_MODEL = 'accounts.User'
+AUTH_USER_MODEL = 'account.User'
 
 
 # Application definition
@@ -44,7 +44,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'accounts.apps.AccountsConfig',
+    'account.apps.AccountConfig',
     'users.apps.UsersConfig',
     'diets.apps.DietsConfig'
 ]
